Crate hexwords

Prints lines from standard input that contain only hexadecimal digits, where ‘a’ through ‘f’ are hexadecimal digits, as well as ‘o’ and ‘i’ (rendered as ‘0’ and ‘1’). For example, here are forty frequent English words of at least length 4 that may be thus represented:

$ sed 's/,.*//' var/unigram_freq.csv | cargo -q run | rg '^.{4}' | head -40 | column
0ff1ce  face    c0ffee  ad0be   cafe    faced   0ecd    bead    c0ca    dec0
c0de    1dea    d1ed    d1ff    1eee    b00b    d1ce    debb1e  ab1de   acad
f00d    feed    dec1ded babe    beef    c1a0    deaf    c0ded   fade    f1fa
added   dead    ac1d    dec1de  decade  edd1e   fe0f    deed    c0c0a   a1ded

Inspired by The Complete List of Hex Words, though I haven’t seen its source code.

Functions

main 🔒

to_hex 🔒